tokendaydata date datatype changed

+++ graphql/queries.js
@@ -769,7 +769,7 @@ const tokendaydatasbydate = {
   async resolve(parent, args, context) {
     try {
       
-      let tokendaydatas = await TokenDayData.find({date:parseInt(args.date)});
+      let tokendaydatas = await TokenDayData.find({date:parseFloat(args.date)});
 
       return tokendaydatas.splice(0, args.first);
     } catch (error) {
